+++ android/src/main/kotlin/io/nekohasekai/sfa/bg/ServiceNotification.kt
@@ -50,7 +50,7 @@ class ServiceNotification(private val status: MutableLiveData<Status>, private v
 
     private val notificationBuilder by lazy {
         NotificationCompat.Builder(service, notificationChannel).setShowWhen(false).setOngoing(true)
-            .setContentTitle("Clash Sing").setOnlyAlertOnce(true)
+            .setContentTitle(PluginManager.applicationLabel).setOnlyAlertOnce(true)
             .setSmallIcon(R.drawable.ic_menu)
             .setCategory(NotificationCompat.CATEGORY_SERVICE)
             .setContentIntent(