+++ example/android/app/src/main/kotlin/com/clashsing/flutter_sing_box_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.clashsing.flutter_sing_box_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()